refactor(#741): refactor accountController
