+++ api/services/AccountService.js
@@ -1,0 +1,8 @@
+const AuthService = require('./AuthService');
+
+module.exports = {
+  getResetPasswordTokenSalt: (user) => {
+    const { dateInscription, id, password } = user;
+    return password + id + dateInscription + AuthService.tokenSalt;
+  },
+};
